Extract createGitignore from create-files
